# frontend/openapi.json
{
  "components": {
    "schemas": {
...
      "BankAccountRead": {
        "description": "Response schema for reading a bank account.\n\n`account_number` is masked to its last 4 digits on the customer-facing\n`routes/bank_accounts.py` endpoints, and returned in full on the admin\nones (`routes/admin/bank_accounts.py`).",
        "properties": {
          "account_holder_name": {
            "title": "Account Holder Name",
            "type": "string"
          },
          "account_number": {
            "title": "Account Number",
            "type": "string"
          },
          "bank_account_id": {
            "format": "uuid",
            "title": "Bank Account Id",
            "type": "string"
          },
          "bank_name": {
            "title": "Bank Name",
            "type": "string"
          },
          "branch_code": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Branch Code"
          },
          "country": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Country"
          },
          "created_at": {
            "title": "Created At",
            "type": "string"
          },
          "currency": {
            "title": "Currency",
            "type": "string"
          },
          "rejection_reason": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Rejection Reason"
          },
          "status": {
            "title": "Status",
            "type": "string"
          }
        },
        "required": [
          "bank_account_id",
          "account_holder_name",
          "bank_name",
          "account_number",
          "branch_code",
          "currency",
          "country",
          "status",
          "rejection_reason",
          "created_at"
        ],
        "title": "BankAccountRead",
        "type": "object"
      },
...
      "CountRead": {
        "properties": {
          "count": {
            "title": "Count",
            "type": "integer"
          }
        },
        "required": [
          "count"
        ],
        "title": "CountRead",
        "type": "object"
      },
...
      "ErrorResponse": {
        "description": "Every refusal that is not a validation failure: ``HTTPException`` and\n``DomainError`` both answer in this shape.",
        "properties": {
          "detail": {
            "title": "Detail",
            "type": "string"
          }
        },
        "required": [
          "detail"
        ],
        "title": "ErrorResponse",
        "type": "object"
      },
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "title": "Detail",
            "type": "array"
          }
        },
        "title": "HTTPValidationError",
        "type": "object"
      },
...
      "RejectBankAccountRequest": {
        "description": "Request schema for rejecting a bank account.",
        "properties": {
          "reason": {
            "title": "Reason",
            "type": "string"
          }
        },
        "required": [
          "reason"
        ],
        "title": "RejectBankAccountRequest",
        "type": "object"
      },
...
      "ValidationError": {
        "properties": {
          "ctx": {
            "title": "Context",
            "type": "object"
          },
          "input": {
            "title": "Input"
          },
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "title": "Location",
            "type": "array"
          },
          "msg": {
            "title": "Message",
            "type": "string"
          },
          "type": {
            "title": "Error Type",
            "type": "string"
          }
        },
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError",
        "type": "object"
      },
...
    },
    "securitySchemes": {
      "ClerkSession": {
        "description": "A Clerk session token.",
        "scheme": "bearer",
        "type": "http"
      }
    }
  },
  "info": {
    "description": "ZAR to RLUSD remittance on the XRP Ledger Testnet. Authenticated routes take a Clerk session token as a bearer token.",
    "title": "RemitX API",
    "version": "0.1.0"
  },
  "openapi": "3.1.0",
  "paths": {
...
    "/admin/bank-accounts/pending": {
      "get": {
        "description": "List every bank account that is `pending_verification` and awaiting an\nadmin's approval or rejection. The full account number is returned here,\nbecause this is an admin endpoint, but the customer-facing endpoints mask\nit for security.",
        "operationId": "list_pending_bank_accounts",
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "items": {
                    "$ref": "#/components/schemas/BankAccountRead"
                  },
                  "title": "Response List Pending Bank Accounts",
                  "type": "array"
                }
              }
            },
            "description": "Successful Response"
          },
          "401": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "No valid session."
          },
          "403": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "The caller lacks the required permission."
          }
        },
        "security": [
          {
            "ClerkSession": []
          }
        ],
        "summary": "List bank accounts awaiting verification",
        "tags": [
          "admin.bank_accounts"
        ]
      }
    },
    "/admin/bank-accounts/pending-count": {
      "get": {
        "description": "How many external bank accounts are `pending_verification`. Needs\n``cashout:read``.",
        "operationId": "get_pending_bank_account_count",
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/CountRead"
                }
              }
            },
            "description": "Successful Response"
          },
          "401": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "No valid session."
          },
          "403": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "The caller lacks the required permission."
          }
        },
        "security": [
          {
            "ClerkSession": []
          }
        ],
        "summary": "Count bank accounts awaiting verification",
        "tags": [
          "admin.bank_accounts"
        ]
      }
    },
    "/admin/bank-accounts/{bank_account_id}/reject": {
      "post": {
        "description": "Reject a user's bank account and record who rejected it and why.",
        "operationId": "reject_bank_account",
        "parameters": [
          {
            "in": "path",
            "name": "bank_account_id",
            "required": true,
            "schema": {
              "format": "uuid",
              "title": "Bank Account Id",
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/RejectBankAccountRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/BankAccountRead"
                }
              }
            },
            "description": "Successful Response"
          },
          "400": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "The request was refused."
          },
          "401": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "No valid session."
          },
          "403": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "The caller lacks the required permission."
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "security": [
          {
            "ClerkSession": []
          }
        ],
        "summary": "Reject a bank account",
        "tags": [
          "admin.bank_accounts"
        ]
      }
    },
    "/admin/bank-accounts/{bank_account_id}/verify": {
      "post": {
        "description": "Verify a user's bank account and record who verified it.",
        "operationId": "verify_bank_account",
        "parameters": [
          {
            "in": "path",
            "name": "bank_account_id",
            "required": true,
            "schema": {
              "format": "uuid",
              "title": "Bank Account Id",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/BankAccountRead"
                }
              }
            },
            "description": "Successful Response"
          },
          "400": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "The request was refused."
          },
          "401": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "No valid session."
          },
          "403": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "The caller lacks the required permission."
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "security": [
          {
            "ClerkSession": []
          }
        ],
        "summary": "Verify a bank account",
        "tags": [
          "admin.bank_accounts"
        ]
      }
    },
...
  },
...
}